% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Collects the MONTH and year associated with a fatal car incident for each year
   provided.}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A vector of years which have an associated file in the current
working directory.}
}
\value{
A list of data frames, each with two columns; MONTH and year. Each
row in a data frame represent a fatal car accident in that month and year.
}
\description{
Collects the MONTH and year associated with a fatal car incident for each year
   provided.
}
\note{
If one of the years specified doesn't have an associated file, and error
   will be thrown.
}
\examples{
years <- fars_read_years(c(2013, 2014, 2015)) 
print(class(years))
[1] "list"

print(years)
[[1]]
# A tibble: 30,202 x 2
   MONTH  year
   <dbl> <dbl>
 1     1  2013
 2     1  2013
 3     1  2013
 4     1  2013
 5     1  2013
 6     1  2013
 7     1  2013
 8     1  2013
 9     1  2013
10     1  2013
# … with 30,192 more rows

[[2]]
# A tibble: 30,056 x 2
   MONTH  year
   <dbl> <dbl>
 1     1  2014
 2     1  2014
 3     1  2014
 4     1  2014
 5     1  2014
 6     1  2014
 7     1  2014
 8     1  2014
 9     1  2014
10     1  2014
# … with 30,046 more rows

[[3]]
# A tibble: 32,166 x 2
   MONTH  year
   <dbl> <dbl>
 1     1  2015
 2     1  2015
 3     1  2015
 4     1  2015
 5     1  2015
 6     1  2015
 7     1  2015
 8     1  2015
 9     1  2015
10     1  2015
# … with 32,156 more rows

}
