% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Creates a table summarizing the count of fatal car incidents in a given month
and year combination.}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{a vector of years for which the number of fatal car incidents
should be displayed, grouped by Month.}
}
\value{
a data frame.
}
\description{
Creates a table summarizing the count of fatal car incidents in a given month
and year combination.
}
\note{
If one of the years specified doesn't have an associated file, and error
   will be thrown.
}
\examples{
summary_df <- fars_summarize_years(c(2013, 2014, 2015)) 
# A tibble: 12 x 4
   MONTH `2013` `2014` `2015`
   <dbl>  <int>  <int>  <int>
 1     1   2230   2168   2368
 2     2   1952   1893   1968
 3     3   2356   2245   2385
 4     4   2300   2308   2430
 5     5   2532   2596   2847
 6     6   2692   2583   2765
 7     7   2660   2696   2998
 8     8   2899   2800   3016
 9     9   2741   2618   2865
10    10   2768   2831   3019
11    11   2615   2714   2724
12    12   2457   2604   2781

}
