% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Collects the MONTH and year associated with a fatal car incident for each year
   provided.}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A vector of years which have an associated file in the current
working directory.}
}
\value{
A list of data frames, each with two columns; MONTH and year. Each
row in a data frame represent a fatal car accident in that month and year.
}
\description{
Collects the MONTH and year associated with a fatal car incident for each year
   provided.
}
\note{
If one of the years specified doesn't have an associated file, and error
   will be thrown.
}
\examples{
\dontrun{years <- fars_read_years(c(2013, 2014, 2015))}

}
